Reflection Serialize & Deserialize 구현 중입니다 ..
 - 기술의 발판인 AnimatorController 완료 .. SerializeByUUID 프로퍼티들 매핑 해주는 일 남음.

+++ DUOLGameEngine_DLL/src/DUOLGameEngine/ECS/Component/MeshCollider.cpp
@@ -72,9 +72,9 @@ namespace DUOLGameEngine
 
 			ResourceManager::GetInstance()->GetMeshInfo(engineMesh, vertices, indices);
 
-			SetVertexBuffer(vertices.data(), vertices.size(), sizeof(DUOLMath::Vector3));
+			SetVertexBuffer(vertices.data(), static_cast<unsigned>(vertices.size()), sizeof(DUOLMath::Vector3));
 
-			SetIndexBuffer(indices.data(), indices.size(), sizeof(UINT32));
+			SetIndexBuffer(indices.data(), static_cast<unsigned>(indices.size()), sizeof(UINT32));
 		}
 	}
 
@@ -85,9 +85,9 @@ namespace DUOLGameEngine
 
 		ResourceManager::GetInstance()->GetMeshInfo(mesh, vertices, indices);
 
-		SetVertexBuffer(vertices.data(), vertices.size(), sizeof(DUOLMath::Vector3));
+		SetVertexBuffer(vertices.data(), static_cast<unsigned>(vertices.size()), sizeof(DUOLMath::Vector3));
 
-		SetIndexBuffer(indices.data(), indices.size(), sizeof(UINT32));
+		SetIndexBuffer(indices.data(), static_cast<unsigned>(indices.size()), sizeof(UINT32));
 	}
 
 	void MeshCollider::SetVertexBuffer(void* buffer, unsigned size, unsigned stride)
